+++ app/tweets/search/page.tsx
@@ -1,57 +1,13 @@
 "use client"
 
-import Footer from "@/app/_components/Footer"
-import Header from "@/app/_components/Header"
-import SearchForm from "@/app/_components/SearchForm"
-import TweetList from "@/app/_components/TweetList"
-import { TweetData } from "@/app/_interfaces/TweetData"
-import { deleteTweet, searchTweets } from "@/app/api/tweets"
-import { useRouter, useSearchParams } from "next/navigation"
-import { useEffect, useState } from "react"
+import SearchPageClient from "@/app/_components/SearchPageClient"
+import { Suspense } from "react"
 
 const SearchPage = () => {
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const query = searchParams.get("query") || ""
-
-  const [tweets, setTweets] = useState<TweetData[]>([])
-
-  useEffect(() => {
-    const fetchTweets = async () => {
-      if (query) {
-        try {
-          const response = await searchTweets(query)
-          setTweets(response)
-        } catch (error) {
-          console.error("ツイートの検索に失敗しました:", error)
-        }
-      }
-    }
-    fetchTweets()
-  }, [query])
-
-  const handleDeleteTweet = async (tweetId: number) => {
-    try {
-      await deleteTweet(tweetId)
-      router.push("/")
-    } catch (error) {
-      console.error("ツイートの削除に失敗しました", error)
-    }
-  }
-
-  const handleSearch = (query: string) => {
-    router.push(`/tweets/search?query=${encodeURIComponent(query)}`)
-  }
-
   return (
-    <>
-      <Header />
-      <SearchForm onSearch={handleSearch} initialQuery={query} />
-      <div className="contents row">
-        <TweetList tweets={tweets} onDeleteTweet={handleDeleteTweet} />
-      </div>
-      <Footer />
-    </>
+    <Suspense fallback={<div>Loading search...</div>}>
+      <SearchPageClient />
+    </Suspense>
   )
 }
